requirements.txt
langchain>=0.0.300
openai>=0.27.0
python-dotenv>=1.0.0
pydantic>=1.10.0
